+++ lib/pymedphys/_mosaiq/mock/__init__.py
@@ -1,1 +1,1 @@
-"""A module for mocking a Mosaiq database"""
+"""A module for creating an MSSQL database that mimics a subset of a MOSAIQ database"""
