+++ lib/pymedphys/_mosaiq/mock/extract_from_bak.py
@@ -4,9 +4,3 @@ By design, only tables and columns that have been defined within
 `data/types_map.toml` are extracted so as to not expose any more of the
 underlying database schema than what is already within the repository.
 """
-
-import pathlib
-
-HERE = pathlib.Path(__file__).parent
-
-TYPES_MAP = HERE / "types_map.toml"
