+++ lib/pymedphys/_mosaiq/mock/paths.py
@@ -1,0 +1,4 @@
+import pathlib
+
+DATA = pathlib.Path(__file__).parent / "data"
+TYPES_MAP = DATA / "types_map.toml"
